pyproject.toml
[tool.poetry]
name = "prueba-backend-fastapi"
version = "0.0.1"
description = ""
authors = ["Álvaro González <45352512+gonz0wsky@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.1"
uvicorn = "^0.18.2"
fastapi = "^0.79.0"
python-dotenv = "^0.20.0"
strawberry-graphql = "^0.130.2"
tortoise-orm = "^0.19.2"
asyncpg = "^0.26.0"
bcrypt = "^3.2.2"
websockets = "^10.3"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
aerich = "^0.6.3"

[tool.aerich]
tortoise_orm = "src.db.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
